+++ model-events/src/main/java/de/focus_shift/urlaubsverwaltung/extension/api/person/PersonDisabledEventDTO.java
@@ -17,7 +17,7 @@ public class PersonDisabledEventDTO {
   @NonNull private UUID id;
   @NonNull private Instant createdAt;
   @NonNull private String tenantId;
-  @NonNull private Integer personId;
+  @NonNull private Long personId;
   @NonNull private String username;
   @NonNull private String lastName;
   @NonNull private String firstName;
@@ -26,7 +26,7 @@ public class PersonDisabledEventDTO {
 
   public static PersonDisabledEventDTO create(
       String tenantId,
-      Integer personId,
+      Long personId,
       String username,
       String lastName,
       String firstName,
@@ -36,7 +36,7 @@ public class PersonDisabledEventDTO {
 
   public static PersonDisabledEventDTO create(
       String tenantId,
-      Integer personId,
+      Long personId,
       String username,
       String lastName,
       String firstName,
